Move responsiblity of maping state to props to ToursList
