+++ app/javascript/packs/application/components/StaffUserDashboard.js
@@ -1,14 +1,8 @@
 import React, { Component } from "react";
 import CreateTourForm from "./CreateTourForm";
 import ToursList from "./ToursList";
-import { connect } from "react-redux";
-import { listTours } from "../redux/tours";
 
 export class StaffUserDashboard extends Component {
-  componentDidMount() {
-    this.props.listTours();
-  }
-
   render() {
     const { username, id, cultural_center } = this.props.currentStaffUser;
     return (
@@ -16,22 +10,10 @@ export class StaffUserDashboard extends Component {
         <h1>Staff Dashboard</h1>
         <h3>Current user: {username}</h3>
         <h3>{cultural_center.name}</h3>
-        <ToursList tours={this.props.tours} />
+        <ToursList />
       </div>
     );
   }
 }
 
-// TODO: add proptypes
-const mapStateToProps = state => ({
-  tours: state.tours
-});
-
-const mapDispatchToProps = { listTours };
-
-const enhance = connect(
-  mapStateToProps,
-  mapDispatchToProps
-);
-
-export default enhance(StaffUserDashboard);
+export default StaffUserDashboard;
